Bugfix: radmc3dData._scalarfieldReader
For binary format data files even if the precision was set to single the
data read was double. Instead of float np.float32 should be used for the
data type. This is now fixed.

+++ radmc3dPy/data.py
@@ -219,7 +219,7 @@ class radmc3dData(object):
                 if hdr[1] == 8:
                     data = np.fromfile(fname, count=-1, dtype=np.float64)
                 elif hdr[1] == 4:
-                    data = np.fromfile(fname, count=-1, dtype=float)
+                    data = np.fromfile(fname, count=-1, dtype=np.float32)
                 else:
                     raise TypeError('Unknown datatype/precision in ' + fname + '. RADMC-3D binary files store 4 byte '
                                     + 'floats or 8 byte doubles. The precision in the file header is '
@@ -245,7 +245,7 @@ class radmc3dData(object):
                 if hdr[1] == 8:
                     data = np.fromfile(fname, count=-1, dtype=np.float64)
                 elif hdr[1] == 4:
-                    data = np.fromfile(fname, count=-1, dtype=float)
+                    data = np.fromfile(fname, count=-1, dtype=np.float32)
                 else:
                     raise TypeError('Unknown datatype/precision in ' + fname + '. RADMC-3D binary files store 4 byte '
                                     + 'floats or 8 byte doubles. The precision in the file header is '
@@ -275,7 +275,6 @@ class radmc3dData(object):
                                      + ' nr of cells in amr_grid.inp : '
                                      + ("%d" % self.grid.nLeaf))
 
-                #data = np.fromfile(fname, count=-1, sep='\n', dtype=np.float64)[3:]
                 data = data.reshape([hdr[1], hdr[2]], order='f')
 
             else:
@@ -288,9 +287,6 @@ class radmc3dData(object):
                                      + ' nr of cells in amr_grid.inp : '
                                      + ("%d" % (self.grid.nx * self.grid.ny * self.grid.nz)))
                 else:
-
-                    # data = np.fromfile(fname, count=-1, sep="\n", dtype=np.float64)
-                    # data = np.fromfile(fname, count=-1, sep="\n", dtype=np.float64)
                     if data.shape[0] == hdr[1] + 2:
                         data = np.reshape(data[2:], [1, self.grid.nz, self.grid.ny, self.grid.nx])
                     elif data.shape[0] == hdr[1] * hdr[2] + 3:
